Fix bug in mm node sorting causing dups. Cache shape coord accesses.

+++ src/algos/mm_router_defs.h
@@ -14,17 +14,15 @@ struct EdgePoint {
   // 0.0 means that the point is at the beginning of the edge, i.e. at
   // fe.from_node_idx. 1.0 means that the start is at the end of the edge. All
   // values between indicate a location between start and end.
-  float to_fraction = 0.0;
+  double to_fraction = 0.0;
   LatLon coord_at_fraction;
-  // LatE6 lat_at_fraction;
-  // LonE6 lon_at_fraction;
   MMFullEdge fe = {};
 
-  float GetFromFraction() const {
-    return std::max(0.0f, std::min(1.0f, 1.0f - to_fraction));
+  double GetFromFraction() const {
+    return std::max(0.0, std::min(1.0, 1.0 - to_fraction));
   }
-  float GetToFraction() const {
-    return std::max(0.0f, std::min(1.0f, to_fraction));
+  double GetToFraction() const {
+    return std::max(0.0, std::min(1.0, to_fraction));
   }
 
   std::string DebugString(const MMCluster& mc, LatE6 origin_lat = {},
@@ -52,15 +50,15 @@ class GeoAnchor {
   GeoAnchor(const LatLon& point) : point_(point) {}
 
   void AddEdge(const EdgePoint& ep) { push_edge(ep); }
-  void AddEdge(const MMCluster& mc, float to_fraction, const MMFullEdge& fe) {
+  void AddEdge(const MMCluster& mc, double to_fraction, const MMFullEdge& fe) {
     // LOG_S(INFO) << "Add edge with fraction: " << fraction;
     push_edge({.to_fraction = to_fraction, .fe = fe});
   }
-  void AddEdge(const MMGraph& mg, float to_fraction, const MMFullEdge& fe) {
+  void AddEdge(const MMGraph& mg, double to_fraction, const MMFullEdge& fe) {
     AddEdge(fe.mc(mg), to_fraction, fe);
   }
 
-  void AddEdge(const MMCluster& mc, float to_fraction, uint32_t from_node_idx,
+  void AddEdge(const MMCluster& mc, double to_fraction, uint32_t from_node_idx,
                uint32_t offset) {
     AddEdge(mc, to_fraction, {from_node_idx, mc.cluster_id, offset});
   }
@@ -236,8 +234,8 @@ struct MMRoutingResult {
     }
     // Handle also the case when there is only one edge, i.e. start is equal to
     // target.
-    float del_frac_start = (fe_pos == 0 ? start.to_fraction : 0.0);
-    float del_frac_target =
+    double del_frac_start = (fe_pos == 0 ? start.to_fraction : 0.0);
+    double del_frac_target =
         (fe_pos == full_edges.size() - 1 ? 1.0 - target.to_fraction : 0.0);
     return std::lround(dist * (1.0 - del_frac_start - del_frac_target));
   }
